GRAND REFACTOR: Major Change in API and DataHelpers.
This is a really really big refactor. Made a lot of changes to the code. And I did not test this well enough.

After using it for a while, and bug fixing on the way, it should become more stable.

But yeah, this is quite important and I really want to finish this refactor.